Common/Core: split out arith prims into minimal primitives.
Things like arithmetic, logical relations and so on will be the same
regardless of language, so having them separate makes it easier to
reuse.
